+++ moderne_cpp/Cia.cpp
@@ -11,7 +11,7 @@ unsigned int Cia::JOU_ChoixCarteAJouer()
 	vector<unique_ptr<CCarte>>& upJOU_mainIA = upJOU_main->PAQ_GetCartes();
 
 	default_random_engine generator(random_device{}());
-	uniform_int_distribution<int> distribution(0, upJOU_mainIA.size() - 1);
+	uniform_int_distribution<unsigned int> distribution(0, (unsigned int)upJOU_mainIA.size() - 1);
 
 	unsigned int iNumeroAleatoire = distribution(generator);
